ON-695 | brien | filter and excludes

+++ migrations/legacy/2019051620191138-content-feed/migration.sh
@@ -16,40 +16,43 @@ else
   printf "No environment specified. Updating environment $http://$1\n"
 fi
 
-newProp="\"byline\": {\"type\": \"nested\", \"dynamic\": \"true\"}";
-
 index=$(curl -X GET "$es:9200/_cat/indices?pretty&s=index:desc" 2>/dev/null | sed -n 's/.*\(published-content[^ \n]*\).*/\1/p' | sed q);
-num=$(echo $index | sed 's/[^0-9]*//g');
-newIndex="$(echo $index | sed 's/[0-9]*//g')$((num+1))";
-alias=$(echo $index | sed 's/\(.*\)_.*/\1/g');
-mappings=$(curl -X GET "$es:9200/$index/_mappings" 2>/dev/null | sed -n "s/\"properties\":{/\"properties\":{$newProp,/p" | sed -n "s/{\"$index\":{\(.*\)}}/\1/p");
+mappings=$(curl -X GET "$es:9200/$index/_mappings");
 
-printf "\n\nCreating new index ($newIndex)...\n\n"
-curl -X GET "$es:9200/$index/_settings" > ./settings.json;
-node ./settings-update.js "$1";
-settings=$(cat ./settings.txt);
-curl -X PUT "$es:9200/$newIndex" -H 'Content-Type: application/json' -d "{$settings,$mappings}";
-rm ./settings.json;
-rm ./settings.txt;
+if [[ $mappings != *"byline"* ]]; then
+    newProp="\"byline\": {\"type\": \"nested\", \"dynamic\": \"true\"}";
+    num=$(echo $index | sed 's/[^0-9]*//g');
+    newIndex="$(echo $index | sed 's/[0-9]*//g')$((num+1))";
+    alias=$(echo $index | sed 's/\(.*\)_.*/\1/g');
+    mappings=$(curl -X GET "$es:9200/$index/_mappings" 2>/dev/null | sed -n "s/\"properties\":{/\"properties\":{$newProp,/p" | sed -n "s/{\"$index\":{\(.*\)}}/\1/p");
 
-printf "\r\n\r\nCopying old index data ($index) to new index ($newIndex)...\n\n"
-curl -X POST "$es:9200/_reindex" -H 'Content-Type: application/json' -d "
-    {
-      \"source\": {
-        \"index\": \"$index\"
-      },
-      \"dest\": {
-        \"index\": \"$newIndex\"
-      }
-    }";
+    printf "\n\nCreating new index ($newIndex)...\n\n"
+    curl -X GET "$es:9200/$index/_settings" > ./settings.json;
+    node ./settings-update.js "$1";
+    settings=$(cat ./settings.txt);
+    curl -X PUT "$es:9200/$newIndex" -H 'Content-Type: application/json' -d "{$settings,$mappings}";
+    rm ./settings.json;
+    rm ./settings.txt;
 
-sleep 1;
+    printf "\r\n\r\nCopying old index data ($index) to new index ($newIndex)...\n\n"
+    curl -X POST "$es:9200/_reindex" -H 'Content-Type: application/json' -d "
+        {
+          \"source\": {
+            \"index\": \"$index\"
+          },
+          \"dest\": {
+            \"index\": \"$newIndex\"
+          }
+        }";
 
-printf "\n\nRemoving old alias and adding new ($alias)...\n\n"
-curl -X POST "$es:9200/_aliases" -H 'Content-Type: application/json' -d "
-    {
-        \"actions\" : [
-            { \"remove\" : { \"index\" : \"$index\", \"alias\" : \"$alias\" } },
-            { \"add\" : { \"index\" : \"$newIndex\", \"alias\" : \"$alias\" } }
-        ]
-    }";
+    sleep 1;
+
+    printf "\n\nRemoving old alias and adding new ($alias)...\n\n"
+    curl -X POST "$es:9200/_aliases" -H 'Content-Type: application/json' -d "
+        {
+            \"actions\" : [
+                { \"remove\" : { \"index\" : \"$index\", \"alias\" : \"$alias\" } },
+                { \"add\" : { \"index\" : \"$newIndex\", \"alias\" : \"$alias\" } }
+            ]
+        }";
+fi
